Fixed index of parameters for tags
According to the parameter to pass to "tags" request:

```
curl https://{subdomain}.zendesk.com/api/v2/tickets/{id}/tags.json \
  -X PUT -d '{ "tags": ["customer"] }' \
  -H "Content-Type: application/json" -v -u {email_address}:{password}
```

what must be passed should be in this format: `{ "tags": ["customer"] }`, but the name of object is singularized by ResourceAbstract.
With this change, I force the object name to not be singularized and it works great now!